services/api-gateway/requirements.txt
fastapi
uvicorn
requests
google-cloud-aiplatform
fpdf2==2.7.9
